+++ Dialogs.m
@@ -18,13 +18,13 @@ http://www.hardcoded.net/licenses/bsd_license
     [a release];
 }
 
-+ (int)askYesNo:(NSString *)message
++ (NSInteger)askYesNo:(NSString *)message
 {
     NSAlert *a = [[NSAlert alloc] init];
     [a addButtonWithTitle:@"Yes"];
     [[a addButtonWithTitle:@"No"] setKeyEquivalent:@"\E"];
     [a setMessageText:message];
-    int r = [a runModal];
+    NSInteger r = [a runModal];
     [a release];
     return r;
 }
